requirements.txt
langchain
langgraph
openai
praw
python-dotenv
streamlit
sentence-transformers
faiss-cpu
numpy
scikit-learn
tiktoken 